returnHitとreturnBlowを同一classに変更
